First pass at retiring the Gateway API
