+++ retired/src/opendap/gateway/Encoder.java
