+++ retired/src/opendap/gateway/HexAsciiEncoder.java
